pin Linux callback delegates to prevent crashes
In a manner akin to that used in the Mac implementation of
VirtualizationInstance, we need to pin our callback delegates
to prevent them from moving in memory; otherwise, upon callback
from the libprojfs C code, we encounter a crash when re-entering
the C# managed system.

See the matching changes in the Mac code added in commit
0d3f0901fb2d59df451157cd525f74f056405771.

We can also remove the NotifyOperationCallback delegate as it is
unused in the Linux code.