use unattended mounts in Linux func tests for now
Until the GVFS.Service process is implemented on Linux,
the "WARNING: ..." message printed when it cannot be contacted
at mount time has the side-effect of confusing the some
functional tests, as the output from the GVFS process is scanned
for any string matching "warning" case-insensitively.

To avoid this, we force the GVFS_UNATTENDED environment variable
to be set on Linux for all functional tests, which in turn
causes the GVFS process to skip checking for a running GVFS.Service
process.